Fixes issues with Infinity, revise types in Array implementation
- Rewrote the subtraction, division and multiplication procedures. There
  were some compliance issues re. Infinity, NaN and 0. Mostly there now.
- Array methods needed to be able to handle arrays of length 2^32
  Replaced a int with unsigned long in a lot of places.
- Maximum number of arguments in a call is now limited to 2^15.

+++ src/runtime/lib/Array.h
@@ -32,9 +32,9 @@ js_val * arr_proto_some(js_val *, js_args *, eval_state *);
 js_val * arr_proto_reduce(js_val *, js_args *, eval_state *);
 js_val * arr_proto_reduce_right(js_val *, js_args *, eval_state *);
 
-void arr_merge_sort(js_prop **, int);
-void arr_merge(js_prop **, js_prop **, int, int, js_prop **);
-void arr_recur(js_prop **, js_prop **, int);
+void arr_merge_sort(js_prop **, unsigned long);
+void arr_merge(js_prop **, js_prop **, unsigned long, unsigned long, js_prop **);
+void arr_recur(js_prop **, js_prop **, unsigned long);
 int arr_cmp(js_prop *, js_prop *);
 int arr_cmp_js(js_prop *, js_prop *);
 int arr_lex_sort(js_prop *, js_prop *);
